refactor(work): move batchload into the Work service transports tree
Relocate pkg/transports/cli/batchload to
pkg/services/work/transports/cli/batchload. The package's only non-test
service import is pkg/services/work, so parking it at the transports root
misrepresented its ownership.

Pure relocation: both source files move unchanged and the two external
importers are updated to the new import path.
